+++ test/prefs/prefs_test.dart
@@ -32,7 +32,7 @@ void main() {
       'instance returns the shared preferences instance',
       () {
         expect(
-          PlayxPrefs.instance,
+          PlayxPrefs.sharedPrefs,
           isA<SharedPreferences>(),
         );
       },